requirements.txt
fastapi
websockets
uvicorn
python-multipart==0.0.18